Add Solutions

+++ C_Simple_Calculator.cpp
@@ -1,12 +1,13 @@
 #include<stdio.h>
 
 int main() {
-    long long a, b;
-    scanf("%lld%lld", &a, &b);
+    int a, b;
+    scanf("%d%d", &a, &b);
 
-    printf("%lld + %lld = %lld\n", a, b, a + b);
-    printf("%lld * %lld = %lld\n", a, b, a * b);
-    printf("%lld - %lld = %lld\n", a, b, a - b);
+    printf("%d + %d = %d\n", a, b, a + b);
+    // printf("%d * %d = %lld\n", a, b, 1LL * a * b);
+    printf("%d * %d = %lld\n", a, b, (long long)a * b);
+    printf("%d - %d = %d\n", a, b, a - b);
     return 0;
 }
